refactor: Ensure integer trial counts and parallel limits in bootstrap functions and update convergence analysis documentation.

+++ +HERA/power_analysis.m
@@ -55,7 +55,7 @@ end
     if isnan(num_simulations) || isinf(num_simulations)
         error('HERA:PowerAnalysis:InvalidSimulations', 'Number of simulations is NaN or Inf. Please check configuration.');
     end % Should be a theoretical Problem therefore not in en.json...
-    num_simulations = round(num_simulations);
+    num_simulations = int32(round(num_simulations));
 
     for m = 1:num_metrics
         % Initialize Win Counters for all pairs
